refactor: move unconstrained.py
The file `unconstrained_layers.py` is moved to the folder `layers/` and
is renamed to `unconstrained.py`.
